+++ app/src/main/cpp/oat-parser/oat/class_status.h
@@ -16,7 +16,7 @@ enum class ClassStatus : uint8_t {
   kSuperclassValidated = 12,  // Superclass validation part of init done.
   kInitializing = 13,  // Class init in progress.
   kInitialized = 14,  // Ready to go.
-  kLast = kInitialized
+  kLast = static_cast<int>(kInitialized)
 };
 std::ostream& operator<<(std::ostream& os, const ClassStatus& rhs);
 }
